pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm_wrapper"
version = "0.1.1"
description = "A unified wrapper for multiple LLM APIs including OpenAI, Gemini, and Anthropic"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Amirsalar Abedini", email = "abediniamirsalar@gmail.com"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
keywords = ["llm", "ai", "openai", "gemini", "anthropic", "wrapper", "api"]

dependencies = [
    "requests>=2.28.0",
    "python-dotenv>=0.20.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "isort>=5.10.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.black]
line-length = 88

[tool.isort]
profile = "black"